[29.10.2020 - Refactoring SideNav Component; commiter: pavel smirnov]
