+++ src/components/dashboard/SideNav/SideNav.tsx
@@ -1,16 +1,27 @@
 /**
  * ********** Импорт основных библиотек из NPM **********
  */
-import React, { useEffect, memo } from 'react';
-import { shallowEqual, useSelector, useDispatch } from 'react-redux';
-import { NavLink } from 'react-router-dom';
+import React, { useEffect, Suspense, lazy, memo } from 'react';
+import { useDispatch } from 'react-redux';
 import { IsOpenSideMenu } from '../../../actions/topNavActions/topNavActions';
-import { PersonalCabinet } from '../../../store/store';
+
+/**
+ * импорт утилит
+ */
+import Loader from '../../../__utils__/Spinner';
 
 /**
  * ********** Импорт файлов стилей **********
  */
 import './SideNav.scss';
+
+/**
+ * импорт компонентов
+ */
+const SideNavMainPart = lazy(() => import('./SideNavMainPart/SideNavMainPart/SideNavMainPart'));
+const SideNavHistoryPart = lazy(() => import('./SideNavHistoryPart/SideNavHistroryPart/SideNavHistoryPart'));
+const SideNavSettingsPart = lazy(() => import('./SideNavSettingsPart/SideNavSettingsPart/SideNavSettingsPart'));
+const SideNavConfiguratorPart = lazy(() => import('./SideNavConfiguratorPart/SideNavConfiguratorPart'));
 
 const SideNav = () => {
     const dispatch = useDispatch();
@@ -32,416 +43,17 @@ const SideNav = () => {
         dispatch(IsOpenSideMenu(isOpen));
     });
 
-    /**
-     * ********** handler обработки клика для раскрытия подменю Главная **********
-     */
-    const handleClickMainSubmenu = (event: React.SyntheticEvent) => {
-        const mainSubMenuElem: Element = event.currentTarget.parentElement!.getElementsByClassName('submenu')[0];
-        const mainSubMenuHandler: Element = event.currentTarget.getElementsByClassName('list-name__icon')[0];
-        mainSubMenuElem.classList.toggle('show');
-        mainSubMenuHandler.classList.toggle('turn');
-    };
-
     return (
-        <nav className="side">
-            <ul className="side__top">
-                <div className="nav-link">
-                    <NavLink to="/dashboard" className="font" onClick={handleClickMainSubmenu}>
-                        <div className="icon">
-                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                <path
-                                    className="cls-1"
-                                    d="M21.63,9.77,11.76.31a1.08,1.08,0,0,0-1.52,0L.37,9.77a1.22,1.22,0,0,0-.29,1.31,1.12,1.12,0,0,0,1.05.76H2.71v9.48a.66.66,0,0,0,.65.68H8.76a.66.66,0,0,0,.65-.68V15.56h3.18v5.76a.66.66,0,0,0,.65.68h5.41a.66.66,0,0,0,.65-.68V11.84h1.58a1.12,1.12,0,0,0,1.05-.76A1.22,1.22,0,0,0,21.63,9.77Z"
-                                />
-                                <path className="cls-1" d="M19.13,1.36H14.78l5,4.78V2A.66.66,0,0,0,19.13,1.36Z" />
-                            </svg>
-                        </div>
-                        <div className="list-name">
-                            <div>Главная</div>
-                            <div className="list-name__icon" />
-                        </div>
-                        <div className="font-block-hover">
-                            <NavLink exact activeClassName="active-page" to="/dashboard" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                        <rect className="cls-1" x="12.39" y="6.74" width="9.61" height="1.34" />
-                                        <polygon
-                                            className="cls-1"
-                                            points="12.39 21.89 15.45 21.89 15.45 15.05 19.03 15.05 19.03 21.89 22 21.89 22 9.41 12.39 9.41 12.39 21.89"
-                                        />
-                                        <path
-                                            className="cls-1"
-                                            d="M0,21.91H11V0H0ZM6.85,2.53H9.61V5.35H6.85Zm0,5.5H9.61v2.82H6.85Zm0,5.52H9.61v2.82H6.84Zm-5.45-11H4.16V5.35H1.39Zm0,5.5H4.16v2.82H1.39Zm0,5.52H4.16v2.82H1.39Z"
-                                        />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Данные о компании</div>
-                            </NavLink>
-                            <NavLink exact activeClassName="active-page" to="/news" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg width="20" height="20" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-                                        <path d="M18.5091 0L11.1182 5.06922H1.70394C1.04366 5.7295 0.660277 6.11289 0 6.77316V11.459C0.660277 12.1193 1.04366 12.5027 1.70394 13.1629H2.9819L6.83706 19.8296C7.05005 19.8935 7.17785 19.9361 7.41214 20L9.66986 18.7859C9.73376 18.5517 9.77636 18.4239 9.84026 18.1896L6.92226 13.1629H11.1395L18.5091 18.2322C18.8072 18.0831 18.8711 18.0405 19.1693 17.8914V0.340788C18.8711 0.191693 18.8072 0.149095 18.5091 0Z" />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Новости</div>
-                            </NavLink>
-                        </div>
-                    </NavLink>
-                    <div className="submenu">
-                        <NavLink exact activeClassName="active-page" to="/dashboard" className="sub-link">
-                            <li className="subparagraph">
-                                <div className="icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                        <rect className="cls-1" x="12.39" y="6.74" width="9.61" height="1.34" />
-                                        <polygon
-                                            className="cls-1"
-                                            points="12.39 21.89 15.45 21.89 15.45 15.05 19.03 15.05 19.03 21.89 22 21.89 22 9.41 12.39 9.41 12.39 21.89"
-                                        />
-                                        <path
-                                            className="cls-1"
-                                            d="M0,21.91H11V0H0ZM6.85,2.53H9.61V5.35H6.85Zm0,5.5H9.61v2.82H6.85Zm0,5.52H9.61v2.82H6.84Zm-5.45-11H4.16V5.35H1.39Zm0,5.5H4.16v2.82H1.39Zm0,5.52H4.16v2.82H1.39Z"
-                                        />
-                                    </svg>
-                                </div>
-                                <div className="list-name">
-                                    <div>Данные о компании</div>
-                                    {/* <div className="list-name__sales">
-                                            <span className="dot" />
-                                            Текущая скидка 22%
-                                            </div>
-                                        */}
-                                    <div className="list-name__icon" />
-                                </div>
-                                <div className="font-block-hover" />
-                            </li>
-                        </NavLink>
-                        <NavLink exact activeClassName="active-page" to="/news" className="sub-link">
-                            <li className="subparagraph">
-                                <div className="icon">
-                                    <svg width="20" height="20" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-                                        <path d="M18.5091 0L11.1182 5.06922H1.70394C1.04366 5.7295 0.660277 6.11289 0 6.77316V11.459C0.660277 12.1193 1.04366 12.5027 1.70394 13.1629H2.9819L6.83706 19.8296C7.05005 19.8935 7.17785 19.9361 7.41214 20L9.66986 18.7859C9.73376 18.5517 9.77636 18.4239 9.84026 18.1896L6.92226 13.1629H11.1395L18.5091 18.2322C18.8072 18.0831 18.8711 18.0405 19.1693 17.8914V0.340788C18.8711 0.191693 18.8072 0.149095 18.5091 0Z" />
-                                    </svg>
-                                </div>
-                                <div className="list-name">
-                                    Новости
-                                    <div className="list-name__icon" />
-                                </div>
-                                <div className="font-block-hover" />
-                            </li>
-                        </NavLink>
-                    </div>
-                </div>
-                {/*
-                    <NavLink exact activeClassName="active-page" to="/price-list" className="nav-link">
-                        <li className="font">
-                            <div className="icon">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20">
-                                    <path d="M10.3324 0H0V20H15.6364V5.304L10.3324 0ZM3.27273 4.36364H6.90909C7.10982 4.36364 7.27273 4.52618 7.27273 4.72727C7.27273 4.92836 7.10982 5.09091 6.90909 5.09091H3.27273C3.072 5.09091 2.90909 4.92836 2.90909 4.72727C2.90909 4.52618 3.072 4.36364 3.27273 4.36364ZM12.3636 16.7273H3.27273C3.072 16.7273 2.90909 16.5647 2.90909 16.3636C2.90909 16.1625 3.072 16 3.27273 16H12.3636C12.5644 16 12.7273 16.1625 12.7273 16.3636C12.7273 16.5647 12.5644 16.7273 12.3636 16.7273ZM12.3636 13.8182H3.27273C3.072 13.8182 2.90909 13.6556 2.90909 13.4545C2.90909 13.2535 3.072 13.0909 3.27273 13.0909H12.3636C12.5644 13.0909 12.7273 13.2535 12.7273 13.4545C12.7273 13.6556 12.5644 13.8182 12.3636 13.8182ZM12.3636 10.9091H3.27273C3.072 10.9091 2.90909 10.7465 2.90909 10.5455C2.90909 10.3444 3.072 10.1818 3.27273 10.1818H12.3636C12.5644 10.1818 12.7273 10.3444 12.7273 10.5455C12.7273 10.7465 12.5644 10.9091 12.3636 10.9091ZM12.3636 8H3.27273C3.072 8 2.90909 7.83745 2.90909 7.63636C2.90909 7.43527 3.072 7.27273 3.27273 7.27273H12.3636C12.5644 7.27273 12.7273 7.43527 12.7273 7.63636C12.7273 7.83745 12.5644 8 12.3636 8ZM9.81818 5.81818V0.727273L14.9091 5.81818H9.81818Z" />
-                                </svg>
-                            </div>
-                            <div className="list-name">Прайс-лист</div>
-                            <div className="font-block-hover">Прайс-лист</div>
-                        </li>
-                    </NavLink>
-                    */}
-                {/*
-                    <NavLink exact activeClassName="active-page" to="/orders" className="nav-link">
-                        <li className="font">
-                            {/** Изменить на subparagraph чтобы вернуть подпараграфы* 
-                            <div className="icon">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20">
-                                    <path d="M7.97772 1.40781C9.06445 1.40781 9.94855 2.29191 9.94855 3.37859H11.3564C11.3564 1.51564 9.84071 0 7.97772 0C6.11477 0 4.59912 1.51564 4.59912 3.37859H6.00693C6.00693 2.29191 6.89103 1.40781 7.97772 1.40781Z" />
-                                    <path d="M15.9513 19.2186L15.0149 4.7329C14.9935 4.3604 14.6853 4.06921 14.3121 4.06921H1.64247C1.26935 4.06921 0.961042 4.3604 0.939691 4.7329L0.00115309 19.256C-0.00992166 19.4493 0.0591547 19.6387 0.192145 19.7795C0.325089 19.9203 0.510216 20 0.703883 20H15.2506C15.251 20 15.2513 20 15.2515 20C15.6403 20 15.9554 19.6849 15.9554 19.2961C15.9555 19.27 15.954 19.2441 15.9513 19.2186ZM6.59246 6.30683C6.59246 6.79895 6.19353 7.19788 5.70141 7.19788C5.20929 7.19788 4.81036 6.79895 4.81036 6.30683V5.79969C4.81036 5.30757 5.20929 4.90864 5.70141 4.90864C6.19353 4.90864 6.59246 5.30757 6.59246 5.79969V6.30683ZM11.1452 6.30683C11.1452 6.79895 10.7462 7.19788 10.2541 7.19788C9.76199 7.19788 9.36307 6.79895 9.36307 6.30683V5.79969C9.36307 5.30757 9.76199 4.90864 10.2541 4.90864C10.7462 4.90864 11.1452 5.30757 11.1452 5.79969V6.30683Z" />
-                                </svg>
-                            </div>
-                            <div className="list-name">Заказы</div>
-                            <div className="font-block-hover">Заказы</div>
-                        </li>
-                    </NavLink>
-                    <NavLink exact activeClassName="active-page" to="/payment" className="nav-link">
-                        <li className="font">
-                            {/** Изменить на subparagraph чтобы вернуть подпараграфы 
-                            <div className="icon">
-                                <svg xmlns="http://www.w3.org/2000/svg">
-                                    <path d="M19.8987 0.10115H1.89831C0.793353 0.10115 -0.101807 0.995989 -0.101807 2.10095V20.1014C-0.101807 21.2063 0.793353 22.1015 1.89831 22.1015H19.8987C21.0037 22.1015 21.8988 21.2063 21.8988 20.1014L21.8985 2.10095C21.8985 0.995989 21.0037 0.10083 19.8984 0.10083L19.8987 0.10115ZM7.14933 5.26327H13.2149C14.9438 5.26327 16.3584 6.67716 16.3584 8.40608V9.26702C16.3584 11.0948 14.8629 12.5902 13.0349 12.5902H9.0212V13.7914H10.342V15.2527H9.0212V17.1245H7.14933V15.2527H5.8704V13.7914H7.14933V12.5902V12.5701H5.87296V11.1088H7.14933V10.9585V5.26327ZM9.03048 6.67908H12.6437C13.6732 6.67908 14.5159 7.52115 14.5159 8.55127V9.06394C14.5159 10.1526 13.6252 11.0436 12.5363 11.0436H9.03048V6.67908Z" />
-                                </svg>
-                            </div>
-                            <div className="list-name">Платежи</div>
-                            <div className="font-block-hover">Платежи</div>
-                        </li>
-                    </NavLink>
-                    <NavLink exact activeClassName="active-page" to="/shipment" className="nav-link">
-                        <li className="font">
-                            {/** Изменить на subparagraph чтобы вернуть подпараграфы 
-                            <div className="icon">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 18">
-                                    <path d="M19.7804 1.07157C19.6339 0.925142 19.4599 0.851685 19.2596 0.851685H7.40754C7.20695 0.851685 7.03324 0.925142 6.88673 1.07157C6.74018 1.2182 6.66684 1.39191 6.66684 1.5925V3.81478H4.81489C4.60652 3.81478 4.38071 3.86493 4.13768 3.96522C3.89461 4.06568 3.69982 4.18892 3.55327 4.33555L1.2615 6.62727C1.16125 6.72757 1.07453 6.84322 1.00108 6.97449C0.927822 7.10543 0.873783 7.2234 0.839041 7.32738C0.80438 7.43165 0.779164 7.57423 0.763881 7.75572C0.748557 7.93714 0.738747 8.07023 0.734896 8.15516C0.731207 8.23996 0.731207 8.38663 0.734896 8.59501C0.738787 8.80346 0.740733 8.93837 0.740733 9.0002V12.7039C0.540186 12.7039 0.366596 12.7771 0.219925 12.9239C0.0733355 13.0702 0 13.244 0 13.4446C0 13.5602 0.0153238 13.6626 0.0462147 13.7512C0.0771462 13.84 0.129158 13.9113 0.202615 13.9654C0.27587 14.0195 0.339557 14.0638 0.393515 14.0986C0.447634 14.1332 0.53828 14.1563 0.665573 14.1679C0.792867 14.1796 0.879621 14.1874 0.925957 14.1912C0.972293 14.1949 1.0706 14.1949 1.22104 14.1912C1.37165 14.1874 1.4584 14.1854 1.48159 14.1854H2.22228C2.22228 15.0031 2.51169 15.7015 3.09027 16.2802C3.66905 16.8589 4.36729 17.1483 5.18525 17.1483C6.00313 17.1483 6.70146 16.8589 7.28012 16.2802C7.8589 15.7015 8.14811 15.0031 8.14811 14.1854H12.5927C12.5927 15.0031 12.8821 15.7015 13.4607 16.2802C14.0393 16.8589 14.7376 17.1483 15.5556 17.1483C16.3737 17.1483 17.0716 16.8589 17.6505 16.2802C18.2292 15.7015 18.5185 15.0031 18.5185 14.1854C18.5415 14.1854 18.6284 14.1874 18.7788 14.1912C18.9293 14.1949 19.0277 14.1949 19.074 14.1912C19.1203 14.1874 19.2073 14.1796 19.3345 14.1679C19.4618 14.1563 19.5523 14.1333 19.6064 14.0986C19.6606 14.0639 19.7241 14.0195 19.7973 13.9654C19.8707 13.9113 19.9229 13.84 19.9537 13.7512C19.9849 13.6625 20 13.5602 20 13.4446V1.59262C20.0002 1.39199 19.9269 1.21848 19.7804 1.07157ZM6.22671 15.2271C5.93365 15.5202 5.58631 15.6671 5.18509 15.6671C4.78383 15.6671 4.43658 15.5203 4.14335 15.2271C3.85013 14.934 3.70359 14.5867 3.70359 14.1855C3.70359 13.7843 3.8503 13.4367 4.14335 13.1436C4.43653 12.8506 4.78379 12.7038 5.18509 12.7038C5.58631 12.7038 5.93361 12.8505 6.22671 13.1436C6.52005 13.4367 6.66668 13.7843 6.66668 14.1855C6.66668 14.5867 6.52005 14.934 6.22671 15.2271ZM6.66668 8.2593H2.22212V7.9122C2.22212 7.81187 2.25678 7.72698 2.32622 7.6575L4.58321 5.40059C4.65261 5.33111 4.73762 5.29649 4.83791 5.29649H6.66668V8.2593ZM16.5972 15.2271C16.3041 15.5202 15.9568 15.6671 15.5556 15.6671C15.1543 15.6671 14.8068 15.5203 14.5137 15.2271C14.2206 14.934 14.0739 14.5867 14.0739 14.1855C14.0739 13.7843 14.2206 13.4367 14.5137 13.1436C14.8068 12.8506 15.1543 12.7038 15.5556 12.7038C15.9568 12.7038 16.3041 12.8505 16.5972 13.1436C16.8903 13.4369 17.0371 13.7843 17.0371 14.1855C17.0371 14.5867 16.8906 14.934 16.5972 15.2271Z" />
-                                </svg>
-                            </div>
-                            <div className="list-name">Отгрузки</div>
-                            <div className="font-block-hover">Отгрузки</div>
-                        </li>
-                    </NavLink>
-                            */}
-                {/* <NavLink exact activeClassName="active-page" to="/profile" className="nav-link">
-                        <li className="font">
-                            <div className="icon">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                    <path d="M11 0C7.68786 0 5.00166 2.5696 5.00166 5.7387V6.6957C5.00166 9.8659 7.68786 12.4344 11.0022 12.4344C14.3165 12.4344 17.0027 9.8659 17.0027 6.6957V5.7387C17.0016 2.5696 14.3154 0 11 0ZM11 15.3043C6.99266 15.3043 1.85236 17.3767 0.373961 19.2159C-0.540139 20.3533 0.329961 22 1.82816 22H20.1718C21.67 22 22.5401 20.3533 21.626 19.217C20.1476 17.3778 15.0051 15.3043 11 15.3043Z" />
-                                </svg>
-                            </div>
-                            <div className="list-name">Профиль</div>
-                            <div className="font-block-hover">Профиль</div>
-                        </li>
-                    </NavLink> 
-                    <NavLink exact activeClassName="active-page" to="/settings" className="nav-link">
-                        <li className="font">
-                            <div className="icon">
-                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                    <path d="M21.4824 8.49506L19.7915 8.12747C19.6444 7.67731 19.4623 7.23822 19.2473 6.81525L20.1834 5.35918C20.3506 5.09901 20.3138 4.75745 20.0953 4.53891L17.4611 1.90472C17.2426 1.68619 16.901 1.64943 16.6408 1.8166L15.1848 2.75269C14.7618 2.53767 14.3227 2.35556 13.8725 2.20853L13.5049 0.517639C13.4393 0.215515 13.1718 0 12.8626 0H9.13741C8.82823 0 8.56068 0.215515 8.49506 0.517639L8.12747 2.20853C7.67731 2.35556 7.23822 2.53767 6.81525 2.75269L5.35918 1.8166C5.09901 1.64943 4.75745 1.68619 4.53891 1.90472L1.90472 4.53891C1.68619 4.75745 1.64943 5.09901 1.8166 5.35918L2.75269 6.81525C2.53767 7.23822 2.35556 7.67731 2.20853 8.12747L0.517639 8.49506C0.215515 8.56085 0 8.82823 0 9.13741V12.8626C0 13.1718 0.215515 13.4391 0.517639 13.5049L2.20853 13.8725C2.35556 14.3227 2.53767 14.7618 2.75269 15.1848L1.8166 16.6408C1.64943 16.901 1.68619 17.2426 1.90472 17.4611L4.53891 20.0953C4.75745 20.3138 5.09901 20.3506 5.35918 20.1834L6.81525 19.2473C7.23822 19.4623 7.67731 19.6444 8.12747 19.7915L8.49506 21.4824C8.56068 21.7845 8.82823 22 9.13741 22H12.8626C13.1718 22 13.4393 21.7845 13.5049 21.4824L13.8725 19.7915C14.3227 19.6444 14.7618 19.4623 15.1848 19.2473L16.6408 20.1834C16.901 20.3506 17.2426 20.314 17.4611 20.0953L20.0953 17.4611C20.3138 17.2426 20.3506 16.901 20.1834 16.6408L19.2473 15.1848C19.4623 14.7618 19.6444 14.3227 19.7915 13.8725L21.4824 13.5049C21.7845 13.4391 22 13.1718 22 12.8626V9.13741C22 8.82823 21.7845 8.56085 21.4824 8.49506ZM14.9442 11C14.9442 13.1748 13.1748 14.9442 11 14.9442C8.82521 14.9442 7.05577 13.1748 7.05577 11C7.05577 8.82521 8.82521 7.05577 11 7.05577C13.1748 7.05577 14.9442 8.82521 14.9442 11Z" />
-                                </svg>
-                            </div>
-                            <div className="list-name">Настройки</div>
-                            <div className="font-block-hover">Настройки</div>
-                        </li>
-                    </NavLink> */}
-                <div className="nav-link">
-                    <NavLink to="/orders" className="font" onClick={handleClickMainSubmenu}>
-                        <div className="icon">
-                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20">
-                                <path d="M10.3324 0H0V20H15.6364V5.304L10.3324 0ZM3.27273 4.36364H6.90909C7.10982 4.36364 7.27273 4.52618 7.27273 4.72727C7.27273 4.92836 7.10982 5.09091 6.90909 5.09091H3.27273C3.072 5.09091 2.90909 4.92836 2.90909 4.72727C2.90909 4.52618 3.072 4.36364 3.27273 4.36364ZM12.3636 16.7273H3.27273C3.072 16.7273 2.90909 16.5647 2.90909 16.3636C2.90909 16.1625 3.072 16 3.27273 16H12.3636C12.5644 16 12.7273 16.1625 12.7273 16.3636C12.7273 16.5647 12.5644 16.7273 12.3636 16.7273ZM12.3636 13.8182H3.27273C3.072 13.8182 2.90909 13.6556 2.90909 13.4545C2.90909 13.2535 3.072 13.0909 3.27273 13.0909H12.3636C12.5644 13.0909 12.7273 13.2535 12.7273 13.4545C12.7273 13.6556 12.5644 13.8182 12.3636 13.8182ZM12.3636 10.9091H3.27273C3.072 10.9091 2.90909 10.7465 2.90909 10.5455C2.90909 10.3444 3.072 10.1818 3.27273 10.1818H12.3636C12.5644 10.1818 12.7273 10.3444 12.7273 10.5455C12.7273 10.7465 12.5644 10.9091 12.3636 10.9091ZM12.3636 8H3.27273C3.072 8 2.90909 7.83745 2.90909 7.63636C2.90909 7.43527 3.072 7.27273 3.27273 7.27273H12.3636C12.5644 7.27273 12.7273 7.43527 12.7273 7.63636C12.7273 7.83745 12.5644 8 12.3636 8ZM9.81818 5.81818V0.727273L14.9091 5.81818H9.81818Z" />
-                            </svg>
-                        </div>
-                        <div className="list-name">
-                            История
-                            <div className="list-name__icon" />
-                        </div>
-                        <div className="font-block-hover">
-                            <NavLink exact activeClassName="active-page" to="/orders" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 18">
-                                        <path d="M7.97772 1.40781C9.06445 1.40781 9.94855 2.29191 9.94855 3.37859H11.3564C11.3564 1.51564 9.84071 0 7.97772 0C6.11477 0 4.59912 1.51564 4.59912 3.37859H6.00693C6.00693 2.29191 6.89103 1.40781 7.97772 1.40781Z" />
-                                        <path d="M15.9513 19.2186L15.0149 4.7329C14.9935 4.3604 14.6853 4.06921 14.3121 4.06921H1.64247C1.26935 4.06921 0.961042 4.3604 0.939691 4.7329L0.00115309 19.256C-0.00992166 19.4493 0.0591547 19.6387 0.192145 19.7795C0.325089 19.9203 0.510216 20 0.703883 20H15.2506C15.251 20 15.2513 20 15.2515 20C15.6403 20 15.9554 19.6849 15.9554 19.2961C15.9555 19.27 15.954 19.2441 15.9513 19.2186ZM6.59246 6.30683C6.59246 6.79895 6.19353 7.19788 5.70141 7.19788C5.20929 7.19788 4.81036 6.79895 4.81036 6.30683V5.79969C4.81036 5.30757 5.20929 4.90864 5.70141 4.90864C6.19353 4.90864 6.59246 5.30757 6.59246 5.79969V6.30683ZM11.1452 6.30683C11.1452 6.79895 10.7462 7.19788 10.2541 7.19788C9.76199 7.19788 9.36307 6.79895 9.36307 6.30683V5.79969C9.36307 5.30757 9.76199 4.90864 10.2541 4.90864C10.7462 4.90864 11.1452 5.30757 11.1452 5.79969V6.30683Z" />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Заказы</div>
-                            </NavLink>
-                            <NavLink exact activeClassName="active-page" to="/payment" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 22">
-                                        <path d="M19.8987 0.10115H1.89831C0.793353 0.10115 -0.101807 0.995989 -0.101807 2.10095V20.1014C-0.101807 21.2063 0.793353 22.1015 1.89831 22.1015H19.8987C21.0037 22.1015 21.8988 21.2063 21.8988 20.1014L21.8985 2.10095C21.8985 0.995989 21.0037 0.10083 19.8984 0.10083L19.8987 0.10115ZM7.14933 5.26327H13.2149C14.9438 5.26327 16.3584 6.67716 16.3584 8.40608V9.26702C16.3584 11.0948 14.8629 12.5902 13.0349 12.5902H9.0212V13.7914H10.342V15.2527H9.0212V17.1245H7.14933V15.2527H5.8704V13.7914H7.14933V12.5902V12.5701H5.87296V11.1088H7.14933V10.9585V5.26327ZM9.03048 6.67908H12.6437C13.6732 6.67908 14.5159 7.52115 14.5159 8.55127V9.06394C14.5159 10.1526 13.6252 11.0436 12.5363 11.0436H9.03048V6.67908Z" />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Платежи</div>
-                            </NavLink>
-                            <NavLink exact activeClassName="active-page" to="/shipment" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 18">
-                                        <path d="M19.7804 1.07157C19.6339 0.925142 19.4599 0.851685 19.2596 0.851685H7.40754C7.20695 0.851685 7.03324 0.925142 6.88673 1.07157C6.74018 1.2182 6.66684 1.39191 6.66684 1.5925V3.81478H4.81489C4.60652 3.81478 4.38071 3.86493 4.13768 3.96522C3.89461 4.06568 3.69982 4.18892 3.55327 4.33555L1.2615 6.62727C1.16125 6.72757 1.07453 6.84322 1.00108 6.97449C0.927822 7.10543 0.873783 7.2234 0.839041 7.32738C0.80438 7.43165 0.779164 7.57423 0.763881 7.75572C0.748557 7.93714 0.738747 8.07023 0.734896 8.15516C0.731207 8.23996 0.731207 8.38663 0.734896 8.59501C0.738787 8.80346 0.740733 8.93837 0.740733 9.0002V12.7039C0.540186 12.7039 0.366596 12.7771 0.219925 12.9239C0.0733355 13.0702 0 13.244 0 13.4446C0 13.5602 0.0153238 13.6626 0.0462147 13.7512C0.0771462 13.84 0.129158 13.9113 0.202615 13.9654C0.27587 14.0195 0.339557 14.0638 0.393515 14.0986C0.447634 14.1332 0.53828 14.1563 0.665573 14.1679C0.792867 14.1796 0.879621 14.1874 0.925957 14.1912C0.972293 14.1949 1.0706 14.1949 1.22104 14.1912C1.37165 14.1874 1.4584 14.1854 1.48159 14.1854H2.22228C2.22228 15.0031 2.51169 15.7015 3.09027 16.2802C3.66905 16.8589 4.36729 17.1483 5.18525 17.1483C6.00313 17.1483 6.70146 16.8589 7.28012 16.2802C7.8589 15.7015 8.14811 15.0031 8.14811 14.1854H12.5927C12.5927 15.0031 12.8821 15.7015 13.4607 16.2802C14.0393 16.8589 14.7376 17.1483 15.5556 17.1483C16.3737 17.1483 17.0716 16.8589 17.6505 16.2802C18.2292 15.7015 18.5185 15.0031 18.5185 14.1854C18.5415 14.1854 18.6284 14.1874 18.7788 14.1912C18.9293 14.1949 19.0277 14.1949 19.074 14.1912C19.1203 14.1874 19.2073 14.1796 19.3345 14.1679C19.4618 14.1563 19.5523 14.1333 19.6064 14.0986C19.6606 14.0639 19.7241 14.0195 19.7973 13.9654C19.8707 13.9113 19.9229 13.84 19.9537 13.7512C19.9849 13.6625 20 13.5602 20 13.4446V1.59262C20.0002 1.39199 19.9269 1.21848 19.7804 1.07157ZM6.22671 15.2271C5.93365 15.5202 5.58631 15.6671 5.18509 15.6671C4.78383 15.6671 4.43658 15.5203 4.14335 15.2271C3.85013 14.934 3.70359 14.5867 3.70359 14.1855C3.70359 13.7843 3.8503 13.4367 4.14335 13.1436C4.43653 12.8506 4.78379 12.7038 5.18509 12.7038C5.58631 12.7038 5.93361 12.8505 6.22671 13.1436C6.52005 13.4367 6.66668 13.7843 6.66668 14.1855C6.66668 14.5867 6.52005 14.934 6.22671 15.2271ZM6.66668 8.2593H2.22212V7.9122C2.22212 7.81187 2.25678 7.72698 2.32622 7.6575L4.58321 5.40059C4.65261 5.33111 4.73762 5.29649 4.83791 5.29649H6.66668V8.2593ZM16.5972 15.2271C16.3041 15.5202 15.9568 15.6671 15.5556 15.6671C15.1543 15.6671 14.8068 15.5203 14.5137 15.2271C14.2206 14.934 14.0739 14.5867 14.0739 14.1855C14.0739 13.7843 14.2206 13.4367 14.5137 13.1436C14.8068 12.8506 15.1543 12.7038 15.5556 12.7038C15.9568 12.7038 16.3041 12.8505 16.5972 13.1436C16.8903 13.4369 17.0371 13.7843 17.0371 14.1855C17.0371 14.5867 16.8906 14.934 16.5972 15.2271Z" />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Отгрузки</div>
-                            </NavLink>
-                        </div>
-                    </NavLink>
-                    <div className="submenu">
-                        <NavLink exact activeClassName="active-page" to="/orders" className="sub-link">
-                            <li className="subparagraph">
-                                <div className="icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20">
-                                        <path d="M7.97772 1.40781C9.06445 1.40781 9.94855 2.29191 9.94855 3.37859H11.3564C11.3564 1.51564 9.84071 0 7.97772 0C6.11477 0 4.59912 1.51564 4.59912 3.37859H6.00693C6.00693 2.29191 6.89103 1.40781 7.97772 1.40781Z" />
-                                        <path d="M15.9513 19.2186L15.0149 4.7329C14.9935 4.3604 14.6853 4.06921 14.3121 4.06921H1.64247C1.26935 4.06921 0.961042 4.3604 0.939691 4.7329L0.00115309 19.256C-0.00992166 19.4493 0.0591547 19.6387 0.192145 19.7795C0.325089 19.9203 0.510216 20 0.703883 20H15.2506C15.251 20 15.2513 20 15.2515 20C15.6403 20 15.9554 19.6849 15.9554 19.2961C15.9555 19.27 15.954 19.2441 15.9513 19.2186ZM6.59246 6.30683C6.59246 6.79895 6.19353 7.19788 5.70141 7.19788C5.20929 7.19788 4.81036 6.79895 4.81036 6.30683V5.79969C4.81036 5.30757 5.20929 4.90864 5.70141 4.90864C6.19353 4.90864 6.59246 5.30757 6.59246 5.79969V6.30683ZM11.1452 6.30683C11.1452 6.79895 10.7462 7.19788 10.2541 7.19788C9.76199 7.19788 9.36307 6.79895 9.36307 6.30683V5.79969C9.36307 5.30757 9.76199 4.90864 10.2541 4.90864C10.7462 4.90864 11.1452 5.30757 11.1452 5.79969V6.30683Z" />
-                                    </svg>
-                                </div>
-                                <div className="list-name">
-                                    Заказы
-                                    <div className="list-name__icon" />
-                                </div>
-                                <div className="font-block-hover">Главная</div>
-                            </li>
-                        </NavLink>
-                        <NavLink exact activeClassName="active-page" to="/payment" className="sub-link">
-                            <li className="subparagraph">
-                                <div className="icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg">
-                                        <path d="M19.8987 0.10115H1.89831C0.793353 0.10115 -0.101807 0.995989 -0.101807 2.10095V20.1014C-0.101807 21.2063 0.793353 22.1015 1.89831 22.1015H19.8987C21.0037 22.1015 21.8988 21.2063 21.8988 20.1014L21.8985 2.10095C21.8985 0.995989 21.0037 0.10083 19.8984 0.10083L19.8987 0.10115ZM7.14933 5.26327H13.2149C14.9438 5.26327 16.3584 6.67716 16.3584 8.40608V9.26702C16.3584 11.0948 14.8629 12.5902 13.0349 12.5902H9.0212V13.7914H10.342V15.2527H9.0212V17.1245H7.14933V15.2527H5.8704V13.7914H7.14933V12.5902V12.5701H5.87296V11.1088H7.14933V10.9585V5.26327ZM9.03048 6.67908H12.6437C13.6732 6.67908 14.5159 7.52115 14.5159 8.55127V9.06394C14.5159 10.1526 13.6252 11.0436 12.5363 11.0436H9.03048V6.67908Z" />
-                                    </svg>
-                                </div>
-                                <div className="list-name">
-                                    Платежи
-                                    <div className="list-name__icon" />
-                                </div>
-                                <div className="font-block-hover">Главная</div>
-                            </li>
-                        </NavLink>
-                        <NavLink exact activeClassName="active-page" to="/shipment" className="sub-link">
-                            <li className="subparagraph">
-                                <div className="icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 18">
-                                        <path d="M19.7804 1.07157C19.6339 0.925142 19.4599 0.851685 19.2596 0.851685H7.40754C7.20695 0.851685 7.03324 0.925142 6.88673 1.07157C6.74018 1.2182 6.66684 1.39191 6.66684 1.5925V3.81478H4.81489C4.60652 3.81478 4.38071 3.86493 4.13768 3.96522C3.89461 4.06568 3.69982 4.18892 3.55327 4.33555L1.2615 6.62727C1.16125 6.72757 1.07453 6.84322 1.00108 6.97449C0.927822 7.10543 0.873783 7.2234 0.839041 7.32738C0.80438 7.43165 0.779164 7.57423 0.763881 7.75572C0.748557 7.93714 0.738747 8.07023 0.734896 8.15516C0.731207 8.23996 0.731207 8.38663 0.734896 8.59501C0.738787 8.80346 0.740733 8.93837 0.740733 9.0002V12.7039C0.540186 12.7039 0.366596 12.7771 0.219925 12.9239C0.0733355 13.0702 0 13.244 0 13.4446C0 13.5602 0.0153238 13.6626 0.0462147 13.7512C0.0771462 13.84 0.129158 13.9113 0.202615 13.9654C0.27587 14.0195 0.339557 14.0638 0.393515 14.0986C0.447634 14.1332 0.53828 14.1563 0.665573 14.1679C0.792867 14.1796 0.879621 14.1874 0.925957 14.1912C0.972293 14.1949 1.0706 14.1949 1.22104 14.1912C1.37165 14.1874 1.4584 14.1854 1.48159 14.1854H2.22228C2.22228 15.0031 2.51169 15.7015 3.09027 16.2802C3.66905 16.8589 4.36729 17.1483 5.18525 17.1483C6.00313 17.1483 6.70146 16.8589 7.28012 16.2802C7.8589 15.7015 8.14811 15.0031 8.14811 14.1854H12.5927C12.5927 15.0031 12.8821 15.7015 13.4607 16.2802C14.0393 16.8589 14.7376 17.1483 15.5556 17.1483C16.3737 17.1483 17.0716 16.8589 17.6505 16.2802C18.2292 15.7015 18.5185 15.0031 18.5185 14.1854C18.5415 14.1854 18.6284 14.1874 18.7788 14.1912C18.9293 14.1949 19.0277 14.1949 19.074 14.1912C19.1203 14.1874 19.2073 14.1796 19.3345 14.1679C19.4618 14.1563 19.5523 14.1333 19.6064 14.0986C19.6606 14.0639 19.7241 14.0195 19.7973 13.9654C19.8707 13.9113 19.9229 13.84 19.9537 13.7512C19.9849 13.6625 20 13.5602 20 13.4446V1.59262C20.0002 1.39199 19.9269 1.21848 19.7804 1.07157ZM6.22671 15.2271C5.93365 15.5202 5.58631 15.6671 5.18509 15.6671C4.78383 15.6671 4.43658 15.5203 4.14335 15.2271C3.85013 14.934 3.70359 14.5867 3.70359 14.1855C3.70359 13.7843 3.8503 13.4367 4.14335 13.1436C4.43653 12.8506 4.78379 12.7038 5.18509 12.7038C5.58631 12.7038 5.93361 12.8505 6.22671 13.1436C6.52005 13.4367 6.66668 13.7843 6.66668 14.1855C6.66668 14.5867 6.52005 14.934 6.22671 15.2271ZM6.66668 8.2593H2.22212V7.9122C2.22212 7.81187 2.25678 7.72698 2.32622 7.6575L4.58321 5.40059C4.65261 5.33111 4.73762 5.29649 4.83791 5.29649H6.66668V8.2593ZM16.5972 15.2271C16.3041 15.5202 15.9568 15.6671 15.5556 15.6671C15.1543 15.6671 14.8068 15.5203 14.5137 15.2271C14.2206 14.934 14.0739 14.5867 14.0739 14.1855C14.0739 13.7843 14.2206 13.4367 14.5137 13.1436C14.8068 12.8506 15.1543 12.7038 15.5556 12.7038C15.9568 12.7038 16.3041 12.8505 16.5972 13.1436C16.8903 13.4369 17.0371 13.7843 17.0371 14.1855C17.0371 14.5867 16.8906 14.934 16.5972 15.2271Z" />
-                                    </svg>
-                                </div>
-                                <div className="list-name">
-                                    Отгрузки
-                                    <div className="list-name__icon" />
-                                </div>
-                                <div className="font-block-hover">Главная</div>
-                            </li>
-                        </NavLink>
-                    </div>
-                </div>
-                {/*
-                <NavLink exact activeClassName="active-page" to="/media" className="nav-link">
-                    <li className="font">
-                        <div className="icon">
-                            <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                <path d="M18.5286 3.33328C18.5286 4.8291 17.3162 6.04172 15.8203 6.04172C14.3245 6.04172 13.1121 4.8291 13.1121 3.33328C13.1121 1.83762 14.3245 0.625 15.8203 0.625C17.3162 0.625 18.5286 1.83762 18.5286 3.33328Z" />
-                                <path d="M15.8203 6.66672C13.982 6.66672 12.4871 5.17166 12.4871 3.33328C12.4871 1.49506 13.982 0 15.8203 0C17.6587 0 19.1536 1.49506 19.1536 3.33328C19.1536 5.17166 17.6587 6.66672 15.8203 6.66672ZM15.8203 1.25C14.6712 1.25 13.7371 2.18506 13.7371 3.33328C13.7371 4.48166 14.6712 5.41672 15.8203 5.41672C16.9695 5.41672 17.9036 4.48166 17.9036 3.33328C17.9036 2.18506 16.9695 1.25 15.8203 1.25Z" />
-                                <path d="M18.5286 16.6667C18.5286 18.1624 17.3162 19.375 15.8203 19.375C14.3245 19.375 13.1121 18.1624 13.1121 16.6667C13.1121 15.1709 14.3245 13.9583 15.8203 13.9583C17.3162 13.9583 18.5286 15.1709 18.5286 16.6667Z" />
-                                <path d="M15.8203 20C13.982 20 12.4871 18.5049 12.4871 16.6667C12.4871 14.8283 13.982 13.3333 15.8203 13.3333C17.6587 13.3333 19.1536 14.8283 19.1536 16.6667C19.1536 18.5049 17.6587 20 15.8203 20ZM15.8203 14.5833C14.6712 14.5833 13.7371 15.5183 13.7371 16.6667C13.7371 17.8149 14.6712 18.75 15.8203 18.75C16.9695 18.75 17.9036 17.8149 17.9036 16.6667C17.9036 15.5183 16.9695 14.5833 15.8203 14.5833Z" />
-                                <path d="M6.86203 10C6.86203 11.4958 5.64941 12.7083 4.15359 12.7083C2.65793 12.7083 1.44531 11.4958 1.44531 10C1.44531 8.50421 2.65793 7.29175 4.15359 7.29175C5.64941 7.29175 6.86203 8.50421 6.86203 10Z" />
-                                <path d="M4.15359 13.3333C2.31537 13.3333 0.820312 11.8384 0.820312 10C0.820312 8.16165 2.31537 6.66675 4.15359 6.66675C5.99197 6.66675 7.48703 8.16165 7.48703 10C7.48703 11.8384 5.99197 13.3333 4.15359 13.3333ZM4.15359 7.91675C3.00446 7.91675 2.07031 8.85165 2.07031 10C2.07031 11.1484 3.00446 12.0833 4.15359 12.0833C5.30289 12.0833 6.23703 11.1484 6.23703 10C6.23703 8.85165 5.30289 7.91675 4.15359 7.91675Z" />
-                                <path d="M6.12121 9.59998C5.83114 9.59998 5.54947 9.44907 5.39612 9.17915C5.16861 8.77998 5.30868 8.27079 5.70785 8.04237L13.4402 3.6341C13.8394 3.40492 14.3486 3.54499 14.577 3.94569C14.8045 4.34486 14.6645 4.85404 14.2653 5.08247L6.53274 9.49073C6.40274 9.56489 6.26114 9.59998 6.12121 9.59998Z" />
-                                <path d="M13.8528 16.4751C13.7127 16.4751 13.5711 16.44 13.4411 16.3658L5.70859 11.9576C5.30942 11.73 5.1695 11.2209 5.39701 10.8208C5.6236 10.4208 6.13355 10.28 6.53363 10.5092L14.2662 14.9175C14.6653 15.145 14.8053 15.6542 14.5777 16.0542C14.4236 16.3242 14.142 16.4751 13.8528 16.4751Z" />
-                            </svg>
-                        </div>
-                        <div className="list-name">Медиа-материалы</div>
-                        <div className="font-block-hover">Медиа-материалы</div>
-                    </li>
-                </NavLink>
-                */}
-                <div className="nav-link">
-                    <NavLink to="/profile" className="font" onClick={handleClickMainSubmenu}>
-                        <div className="icon">
-                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                <path d="M21.4824 8.49506L19.7915 8.12747C19.6444 7.67731 19.4623 7.23822 19.2473 6.81525L20.1834 5.35918C20.3506 5.09901 20.3138 4.75745 20.0953 4.53891L17.4611 1.90472C17.2426 1.68619 16.901 1.64943 16.6408 1.8166L15.1848 2.75269C14.7618 2.53767 14.3227 2.35556 13.8725 2.20853L13.5049 0.517639C13.4393 0.215515 13.1718 0 12.8626 0H9.13741C8.82823 0 8.56068 0.215515 8.49506 0.517639L8.12747 2.20853C7.67731 2.35556 7.23822 2.53767 6.81525 2.75269L5.35918 1.8166C5.09901 1.64943 4.75745 1.68619 4.53891 1.90472L1.90472 4.53891C1.68619 4.75745 1.64943 5.09901 1.8166 5.35918L2.75269 6.81525C2.53767 7.23822 2.35556 7.67731 2.20853 8.12747L0.517639 8.49506C0.215515 8.56085 0 8.82823 0 9.13741V12.8626C0 13.1718 0.215515 13.4391 0.517639 13.5049L2.20853 13.8725C2.35556 14.3227 2.53767 14.7618 2.75269 15.1848L1.8166 16.6408C1.64943 16.901 1.68619 17.2426 1.90472 17.4611L4.53891 20.0953C4.75745 20.3138 5.09901 20.3506 5.35918 20.1834L6.81525 19.2473C7.23822 19.4623 7.67731 19.6444 8.12747 19.7915L8.49506 21.4824C8.56068 21.7845 8.82823 22 9.13741 22H12.8626C13.1718 22 13.4393 21.7845 13.5049 21.4824L13.8725 19.7915C14.3227 19.6444 14.7618 19.4623 15.1848 19.2473L16.6408 20.1834C16.901 20.3506 17.2426 20.314 17.4611 20.0953L20.0953 17.4611C20.3138 17.2426 20.3506 16.901 20.1834 16.6408L19.2473 15.1848C19.4623 14.7618 19.6444 14.3227 19.7915 13.8725L21.4824 13.5049C21.7845 13.4391 22 13.1718 22 12.8626V9.13741C22 8.82823 21.7845 8.56085 21.4824 8.49506ZM14.9442 11C14.9442 13.1748 13.1748 14.9442 11 14.9442C8.82521 14.9442 7.05577 13.1748 7.05577 11C7.05577 8.82521 8.82521 7.05577 11 7.05577C13.1748 7.05577 14.9442 8.82521 14.9442 11Z" />
-                            </svg>
-                        </div>
-                        <div className="list-name">
-                            Настройки
-                            <div className="list-name__icon" />
-                        </div>
-                        <div className="font-block-hover">
-                            <NavLink exact activeClassName="active-page" to="/profile" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                        <path d="M11 0C7.68786 0 5.00166 2.5696 5.00166 5.7387V6.6957C5.00166 9.8659 7.68786 12.4344 11.0022 12.4344C14.3165 12.4344 17.0027 9.8659 17.0027 6.6957V5.7387C17.0016 2.5696 14.3154 0 11 0ZM11 15.3043C6.99266 15.3043 1.85236 17.3767 0.373961 19.2159C-0.540139 20.3533 0.329961 22 1.82816 22H20.1718C21.67 22 22.5401 20.3533 21.626 19.217C20.1476 17.3778 15.0051 15.3043 11 15.3043Z" />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Профиль</div>
-                            </NavLink>
-                            <NavLink exact activeClassName="active-page" to="/users-and-roles" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg width="20" height="20" viewBox="0 0 20 20" fill="black" xmlns="http://www.w3.org/2000/svg">
-                                        <path d="M8.72081 0.97044C9.12344 -0.32348 10.8766 -0.32348 11.2792 0.970441L12.7178 5.59356C12.8979 6.17221 13.4143 6.564 13.997 6.564H18.6524C19.9554 6.564 20.4971 8.30496 19.443 9.10464L15.6767 11.9619C15.2053 12.3195 15.008 12.9534 15.1881 13.5321L16.6267 18.1552C17.0293 19.4491 15.611 20.5251 14.5569 19.7254L10.7906 16.8682C10.3192 16.5105 9.68083 16.5105 9.20941 16.8682L5.44311 19.7254C4.389 20.5251 2.97069 19.4491 3.37333 18.1552L4.81193 13.5321C4.99199 12.9534 4.79473 12.3195 4.32332 11.9619L0.557018 9.10464C-0.497098 8.30495 0.0446463 6.564 1.3476 6.564H6.00301C6.58571 6.564 7.10214 6.17221 7.28221 5.59356L8.72081 0.97044Z" />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Роли и пользователи</div>
-                            </NavLink>
-                        </div>
-                    </NavLink>
-                    <div className="submenu">
-                        <NavLink exact activeClassName="active-page" to="/profile" className="sub-link">
-                            <li className="subparagraph">
-                                <div className="icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                        <path d="M11 0C7.68786 0 5.00166 2.5696 5.00166 5.7387V6.6957C5.00166 9.8659 7.68786 12.4344 11.0022 12.4344C14.3165 12.4344 17.0027 9.8659 17.0027 6.6957V5.7387C17.0016 2.5696 14.3154 0 11 0ZM11 15.3043C6.99266 15.3043 1.85236 17.3767 0.373961 19.2159C-0.540139 20.3533 0.329961 22 1.82816 22H20.1718C21.67 22 22.5401 20.3533 21.626 19.217C20.1476 17.3778 15.0051 15.3043 11 15.3043Z" />
-                                    </svg>
-                                </div>
-                                <div className="list-name">
-                                    Профиль
-                                    <div className="list-name__icon" />
-                                </div>
-                                <div className="font-block-hover">Главная</div>
-                            </li>
-                        </NavLink>
-                        <NavLink exact activeClassName="active-page" to="/users-and-roles" className="sub-link">
-                            <li className="subparagraph">
-                                <div className="icon">
-                                    <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                        <path d="M8.72081 0.97044C9.12344 -0.32348 10.8766 -0.32348 11.2792 0.970441L12.7178 5.59356C12.8979 6.17221 13.4143 6.564 13.997 6.564H18.6524C19.9554 6.564 20.4971 8.30496 19.443 9.10464L15.6767 11.9619C15.2053 12.3195 15.008 12.9534 15.1881 13.5321L16.6267 18.1552C17.0293 19.4491 15.611 20.5251 14.5569 19.7254L10.7906 16.8682C10.3192 16.5105 9.68083 16.5105 9.20941 16.8682L5.44311 19.7254C4.389 20.5251 2.97069 19.4491 3.37333 18.1552L4.81193 13.5321C4.99199 12.9534 4.79473 12.3195 4.32332 11.9619L0.557018 9.10464C-0.497098 8.30495 0.0446463 6.564 1.3476 6.564H6.00301C6.58571 6.564 7.10214 6.17221 7.28221 5.59356L8.72081 0.97044Z" />
-                                    </svg>
-                                </div>
-                                <div className="list-name">
-                                    Роли и пользователи
-                                    <div className="list-name__icon" />
-                                </div>
-                                <div className="font-block-hover">Главная</div>
-                            </li>
-                        </NavLink>
-                        {/*
-                            <NavLink exact activeClassName="active-page" to="/notification" className="sub-link">
-                                <li className="subparagraph">
-                                    <div className="icon">
-                                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 19.21 22">
-                                            <path
-                                                className="cls-1"
-                                                d="M4.62,4.25a7,7,0,0,0-2.06,5V11.7a10.47,10.47,0,0,1-.82,4.05H17.46a10.49,10.49,0,0,1-.82-4.05V9.22a7,7,0,0,0-5.4-6.84A1.76,1.76,0,1,0,8,2.39,7,7,0,0,0,4.62,4.25ZM.26,17.31a.88.88,0,0,0,.63,1.5H7.64a2.2,2.2,0,1,0,3.92,0h6.75a.88.88,0,1,0,0-1.76H.89a.88.88,0,0,0-.63.26Z"
-                                            />
-                                        </svg>
-                                    </div>
-                                    <div className="list-name">
-                                        Уведомления
-                                        <div className="list-name__icon" />
-                                    </div>
-                                    <div className="font-block-hover">Главная</div>
-                                </li>
-                            </NavLink>
-                            */}
-                    </div>
-                </div>
-                <div className="nav-link">
-                    <NavLink activeClassName="active-page" to="/configurator" className="font">
-                        <div className="icon">
-                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                <path
-                                    className="cls-1"
-                                    d="M5,12.36H17a4.94,4.94,0,0,1,5,4.82A4.94,4.94,0,0,1,17,22H5a4.94,4.94,0,0,1-5-4.82A4.94,4.94,0,0,1,5,12.36ZM17,20.68a3.51,3.51,0,1,0,0-7H5a3.51,3.51,0,1,0,0,7Z"
-                                />
-                                <path
-                                    className="cls-1"
-                                    d="M17.18,15A2.16,2.16,0,1,1,15,17.18,2.16,2.16,0,0,1,17.18,15Zm0,3a.86.86,0,1,0-.86-.86A.86.86,0,0,0,17.18,18Z"
-                                />
-                                <path
-                                    className="cls-1"
-                                    d="M17,9.64H5A4.94,4.94,0,0,1,0,4.82,4.94,4.94,0,0,1,5,0H17a4.94,4.94,0,0,1,5,4.82A4.94,4.94,0,0,1,17,9.64ZM5,1.32A3.59,3.59,0,0,0,1.38,4.82,3.59,3.59,0,0,0,5,8.33H17a3.59,3.59,0,0,0,3.67-3.51A3.59,3.59,0,0,0,17,1.32Z"
-                                />
-                                <path
-                                    className="cls-1"
-                                    d="M4.82,7A2.16,2.16,0,1,1,7,4.82,2.16,2.16,0,0,1,4.82,7Zm0-3a.86.86,0,1,0,.86.86A.86.86,0,0,0,4.82,4Z"
-                                />
-                            </svg>
-                        </div>
-                        <div className="list-name">Конфигуратор</div>
-                        <div className="font-block-hover">
-                            <NavLink activeClassName="active-page" to="/configurator" className="hover__block">
-                                <div className="hover__block--icon">
-                                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 22 22">
-                                        <path
-                                            className="cls-1"
-                                            d="M5,12.36H17a4.94,4.94,0,0,1,5,4.82A4.94,4.94,0,0,1,17,22H5a4.94,4.94,0,0,1-5-4.82A4.94,4.94,0,0,1,5,12.36ZM17,20.68a3.51,3.51,0,1,0,0-7H5a3.51,3.51,0,1,0,0,7Z"
-                                        />
-                                        <path
-                                            className="cls-1"
-                                            d="M17.18,15A2.16,2.16,0,1,1,15,17.18,2.16,2.16,0,0,1,17.18,15Zm0,3a.86.86,0,1,0-.86-.86A.86.86,0,0,0,17.18,18Z"
-                                        />
-                                        <path
-                                            className="cls-1"
-                                            d="M17,9.64H5A4.94,4.94,0,0,1,0,4.82,4.94,4.94,0,0,1,5,0H17a4.94,4.94,0,0,1,5,4.82A4.94,4.94,0,0,1,17,9.64ZM5,1.32A3.59,3.59,0,0,0,1.38,4.82,3.59,3.59,0,0,0,5,8.33H17a3.59,3.59,0,0,0,3.67-3.51A3.59,3.59,0,0,0,17,1.32Z"
-                                        />
-                                        <path
-                                            className="cls-1"
-                                            d="M4.82,7A2.16,2.16,0,1,1,7,4.82,2.16,2.16,0,0,1,4.82,7Zm0-3a.86.86,0,1,0,.86.86A.86.86,0,0,0,4.82,4Z"
-                                        />
-                                    </svg>
-                                </div>
-                                <div className="hover__block--text">Конфигуратор</div>
-                            </NavLink>
-                        </div>
-                    </NavLink>
-                </div>
-            </ul>
-        </nav>
+        <Suspense fallback={Loader}>
+            <nav className="side">
+                <ul className="side__top">
+                    <SideNavMainPart />
+                    <SideNavHistoryPart />
+                    <SideNavSettingsPart />
+                    <SideNavConfiguratorPart />
+                </ul>
+            </nav>
+        </Suspense>
     );
 };
