+++ src/components/dashboard/SideNav/SideNavHistoryPart/SideNavHistoryPartSubmenu/SideNavHistoryPartSubmenu.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+
+const SideNavHistoryPartSubmenu = () => {
+    return (
+        <div className="submenu">
+            <NavLink exact activeClassName="active-page" to="/orders" className="sub-link">
+                <li className="subparagraph">
+                    <div className="icon">
+                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 20">
+                            <path d="M7.97772 1.40781C9.06445 1.40781 9.94855 2.29191 9.94855 3.37859H11.3564C11.3564 1.51564 9.84071 0 7.97772 0C6.11477 0 4.59912 1.51564 4.59912 3.37859H6.00693C6.00693 2.29191 6.89103 1.40781 7.97772 1.40781Z" />
+                            <path d="M15.9513 19.2186L15.0149 4.7329C14.9935 4.3604 14.6853 4.06921 14.3121 4.06921H1.64247C1.26935 4.06921 0.961042 4.3604 0.939691 4.7329L0.00115309 19.256C-0.00992166 19.4493 0.0591547 19.6387 0.192145 19.7795C0.325089 19.9203 0.510216 20 0.703883 20H15.2506C15.251 20 15.2513 20 15.2515 20C15.6403 20 15.9554 19.6849 15.9554 19.2961C15.9555 19.27 15.954 19.2441 15.9513 19.2186ZM6.59246 6.30683C6.59246 6.79895 6.19353 7.19788 5.70141 7.19788C5.20929 7.19788 4.81036 6.79895 4.81036 6.30683V5.79969C4.81036 5.30757 5.20929 4.90864 5.70141 4.90864C6.19353 4.90864 6.59246 5.30757 6.59246 5.79969V6.30683ZM11.1452 6.30683C11.1452 6.79895 10.7462 7.19788 10.2541 7.19788C9.76199 7.19788 9.36307 6.79895 9.36307 6.30683V5.79969C9.36307 5.30757 9.76199 4.90864 10.2541 4.90864C10.7462 4.90864 11.1452 5.30757 11.1452 5.79969V6.30683Z" />
+                        </svg>
+                    </div>
+                    <div className="list-name">
+                        Заказы
+                        <div className="list-name__icon" />
+                    </div>
+                    <div className="font-block-hover">Главная</div>
+                </li>
+            </NavLink>
+            <NavLink exact activeClassName="active-page" to="/payment" className="sub-link">
+                <li className="subparagraph">
+                    <div className="icon">
+                        <svg xmlns="http://www.w3.org/2000/svg">
+                            <path d="M19.8987 0.10115H1.89831C0.793353 0.10115 -0.101807 0.995989 -0.101807 2.10095V20.1014C-0.101807 21.2063 0.793353 22.1015 1.89831 22.1015H19.8987C21.0037 22.1015 21.8988 21.2063 21.8988 20.1014L21.8985 2.10095C21.8985 0.995989 21.0037 0.10083 19.8984 0.10083L19.8987 0.10115ZM7.14933 5.26327H13.2149C14.9438 5.26327 16.3584 6.67716 16.3584 8.40608V9.26702C16.3584 11.0948 14.8629 12.5902 13.0349 12.5902H9.0212V13.7914H10.342V15.2527H9.0212V17.1245H7.14933V15.2527H5.8704V13.7914H7.14933V12.5902V12.5701H5.87296V11.1088H7.14933V10.9585V5.26327ZM9.03048 6.67908H12.6437C13.6732 6.67908 14.5159 7.52115 14.5159 8.55127V9.06394C14.5159 10.1526 13.6252 11.0436 12.5363 11.0436H9.03048V6.67908Z" />
+                        </svg>
+                    </div>
+                    <div className="list-name">
+                        Платежи
+                        <div className="list-name__icon" />
+                    </div>
+                    <div className="font-block-hover">Главная</div>
+                </li>
+            </NavLink>
+            <NavLink exact activeClassName="active-page" to="/shipment" className="sub-link">
+                <li className="subparagraph">
+                    <div className="icon">
+                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 18">
+                            <path d="M19.7804 1.07157C19.6339 0.925142 19.4599 0.851685 19.2596 0.851685H7.40754C7.20695 0.851685 7.03324 0.925142 6.88673 1.07157C6.74018 1.2182 6.66684 1.39191 6.66684 1.5925V3.81478H4.81489C4.60652 3.81478 4.38071 3.86493 4.13768 3.96522C3.89461 4.06568 3.69982 4.18892 3.55327 4.33555L1.2615 6.62727C1.16125 6.72757 1.07453 6.84322 1.00108 6.97449C0.927822 7.10543 0.873783 7.2234 0.839041 7.32738C0.80438 7.43165 0.779164 7.57423 0.763881 7.75572C0.748557 7.93714 0.738747 8.07023 0.734896 8.15516C0.731207 8.23996 0.731207 8.38663 0.734896 8.59501C0.738787 8.80346 0.740733 8.93837 0.740733 9.0002V12.7039C0.540186 12.7039 0.366596 12.7771 0.219925 12.9239C0.0733355 13.0702 0 13.244 0 13.4446C0 13.5602 0.0153238 13.6626 0.0462147 13.7512C0.0771462 13.84 0.129158 13.9113 0.202615 13.9654C0.27587 14.0195 0.339557 14.0638 0.393515 14.0986C0.447634 14.1332 0.53828 14.1563 0.665573 14.1679C0.792867 14.1796 0.879621 14.1874 0.925957 14.1912C0.972293 14.1949 1.0706 14.1949 1.22104 14.1912C1.37165 14.1874 1.4584 14.1854 1.48159 14.1854H2.22228C2.22228 15.0031 2.51169 15.7015 3.09027 16.2802C3.66905 16.8589 4.36729 17.1483 5.18525 17.1483C6.00313 17.1483 6.70146 16.8589 7.28012 16.2802C7.8589 15.7015 8.14811 15.0031 8.14811 14.1854H12.5927C12.5927 15.0031 12.8821 15.7015 13.4607 16.2802C14.0393 16.8589 14.7376 17.1483 15.5556 17.1483C16.3737 17.1483 17.0716 16.8589 17.6505 16.2802C18.2292 15.7015 18.5185 15.0031 18.5185 14.1854C18.5415 14.1854 18.6284 14.1874 18.7788 14.1912C18.9293 14.1949 19.0277 14.1949 19.074 14.1912C19.1203 14.1874 19.2073 14.1796 19.3345 14.1679C19.4618 14.1563 19.5523 14.1333 19.6064 14.0986C19.6606 14.0639 19.7241 14.0195 19.7973 13.9654C19.8707 13.9113 19.9229 13.84 19.9537 13.7512C19.9849 13.6625 20 13.5602 20 13.4446V1.59262C20.0002 1.39199 19.9269 1.21848 19.7804 1.07157ZM6.22671 15.2271C5.93365 15.5202 5.58631 15.6671 5.18509 15.6671C4.78383 15.6671 4.43658 15.5203 4.14335 15.2271C3.85013 14.934 3.70359 14.5867 3.70359 14.1855C3.70359 13.7843 3.8503 13.4367 4.14335 13.1436C4.43653 12.8506 4.78379 12.7038 5.18509 12.7038C5.58631 12.7038 5.93361 12.8505 6.22671 13.1436C6.52005 13.4367 6.66668 13.7843 6.66668 14.1855C6.66668 14.5867 6.52005 14.934 6.22671 15.2271ZM6.66668 8.2593H2.22212V7.9122C2.22212 7.81187 2.25678 7.72698 2.32622 7.6575L4.58321 5.40059C4.65261 5.33111 4.73762 5.29649 4.83791 5.29649H6.66668V8.2593ZM16.5972 15.2271C16.3041 15.5202 15.9568 15.6671 15.5556 15.6671C15.1543 15.6671 14.8068 15.5203 14.5137 15.2271C14.2206 14.934 14.0739 14.5867 14.0739 14.1855C14.0739 13.7843 14.2206 13.4367 14.5137 13.1436C14.8068 12.8506 15.1543 12.7038 15.5556 12.7038C15.9568 12.7038 16.3041 12.8505 16.5972 13.1436C16.8903 13.4369 17.0371 13.7843 17.0371 14.1855C17.0371 14.5867 16.8906 14.934 16.5972 15.2271Z" />
+                        </svg>
+                    </div>
+                    <div className="list-name">
+                        Отгрузки
+                        <div className="list-name__icon" />
+                    </div>
+                    <div className="font-block-hover">Главная</div>
+                </li>
+            </NavLink>
+        </div>
+    );
+};
+
+export default SideNavHistoryPartSubmenu;
